// src/gen/steps.ts
export interface SearchHit {
	/**
	 * The parser attempts to read fields 1-5 as either string or varint.
	 * We represent them as generic fields here.
	 * These likely correspond to file path, line number, matched text, etc.
	 */
	field1: string;
	field2: string;
	field3: string;
	field4: string;
	field5: string;
}
// ...
export const SearchHit: MessageFns<SearchHit> = {
	encode(
		message: SearchHit,
		writer: BinaryWriter = new BinaryWriter(),
	): BinaryWriter {
		if (message.field1 !== "") {
			writer.uint32(10).string(message.field1);
		}
		if (message.field2 !== "") {
			writer.uint32(18).string(message.field2);
		}
		if (message.field3 !== "") {
			writer.uint32(26).string(message.field3);
		}
		if (message.field4 !== "") {
			writer.uint32(34).string(message.field4);
		}
		if (message.field5 !== "") {
			writer.uint32(42).string(message.field5);
		}
		return writer;
	},

	decode(input: BinaryReader | Uint8Array, length?: number): SearchHit {
		const reader =
			input instanceof BinaryReader ? input : new BinaryReader(input);
		const end = length === undefined ? reader.len : reader.pos + length;
		const message = createBaseSearchHit();
		while (reader.pos < end) {
			const tag = reader.uint32();
			switch (tag >>> 3) {
				case 1: {
					if (tag !== 10) {
						break;
					}

					message.field1 = reader.string();
					continue;
				}
				case 2: {
					if (tag !== 18) {
						break;
					}

					message.field2 = reader.string();
					continue;
				}
				case 3: {
					if (tag !== 26) {
						break;
					}

					message.field3 = reader.string();
					continue;
				}
				case 4: {
					if (tag !== 34) {
						break;
					}

					message.field4 = reader.string();
					continue;
				}
				case 5: {
					if (tag !== 42) {
						break;
					}

					message.field5 = reader.string();
					continue;
				}
			}
			if ((tag & 7) === 4 || tag === 0) {
				break;
			}
			reader.skip(tag & 7);
		}
		return message;
	},

	fromJSON(object: any): SearchHit {
		return {
			field1: isSet(object.field1) ? globalThis.String(object.field1) : "",
			field2: isSet(object.field2) ? globalThis.String(object.field2) : "",
			field3: isSet(object.field3) ? globalThis.String(object.field3) : "",
			field4: isSet(object.field4) ? globalThis.String(object.field4) : "",
			field5: isSet(object.field5) ? globalThis.String(object.field5) : "",
		};
	},

	toJSON(message: SearchHit): unknown {
		const obj: any = {};
		if (message.field1 !== "") {
			obj.field1 = message.field1;
		}
		if (message.field2 !== "") {
			obj.field2 = message.field2;
		}
		if (message.field3 !== "") {
			obj.field3 = message.field3;
		}
		if (message.field4 !== "") {
			obj.field4 = message.field4;
		}
		if (message.field5 !== "") {
			obj.field5 = message.field5;
		}
		return obj;
	},

	create<I extends Exact<DeepPartial<SearchHit>, I>>(base?: I): SearchHit {
		return SearchHit.fromPartial(base ?? ({} as any));
	},
	fromPartial<I extends Exact<DeepPartial<SearchHit>, I>>(
		object: I,
	): SearchHit {
		const message = createBaseSearchHit();
		message.field1 = object.field1 ?? "";
		message.field2 = object.field2 ?? "";
		message.field3 = object.field3 ?? "";
		message.field4 = object.field4 ?? "";
		message.field5 = object.field5 ?? "";
		return message;
	},
};
// ...
type Builtin =
	| Date
	| Function
	| Uint8Array
	| string
	| number
	| boolean
	| undefined;

export type DeepPartial<T> = T extends Builtin
	? T
	: T extends globalThis.Array<infer U>
		? globalThis.Array<DeepPartial<U>>
		: T extends ReadonlyArray<infer U>
			? ReadonlyArray<DeepPartial<U>>
			: T extends {}
				? { [K in keyof T]?: DeepPartial<T[K]> }
				: Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin
	? P
	: P & { [K in keyof P]: Exact<P[K], I[K]> } & {
			[K in Exclude<keyof I, KeysOfUnion<P>>]: never;
		};
// ...
export interface MessageFns<T> {
	encode(message: T, writer?: BinaryWriter): BinaryWriter;
	decode(input: BinaryReader | Uint8Array, length?: number): T;
	fromJSON(object: any): T;
	toJSON(message: T): unknown;
	create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
	fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}